requirements.txt
streamlit
sentence-transformers
faiss-cpu
numpy
pandas
scikit-learn
plotly
datasets
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
pyarrow
